requirements.txt
langchain-groq==0.1.3
langchain-community==0.0.38
langchain-core==0.1.52
python-dotenv==1.0.1
streamlit==1.34.0
pydantic==2.7.1